chore(root): Introduce set-package-dependencies.mjs script
Manipulate novu dependencies per use case:
1. Either restore the workspace:* protocol after a release
2. Or set them to latest version before an NPM package preview deployment powered by pkg.pr.new
